+++ R/plot.arrivals.R
@@ -1,21 +1,30 @@
-plot_arrivals <- function(x, metric=c("activity_time", "waiting_time", "flow_time"), ...) {
+#' @name get_mon
+#' @export
+get_mon_arrivals <- function(...) {
+  x <- simmer::get_mon_arrivals(...)
+  class(x) <- c("arrivals", class(x))
+  x
+}
+
+#' @name plot.mon
+#' @export
+plot.arrivals <- function(x, metric=c("activity_time", "waiting_time", "flow_time"), ...) {
   metric <- match.arg(metric)
 
-  monitor_data <- get_mon_arrivals(x)
-  if (nrow(monitor_data) == 0)
+  if (nrow(x) == 0)
     stop("no data available")
 
-  monitor_data <- monitor_data %>%
+  x <- x %>%
     dplyr::mutate(flow_time = .data$end_time - .data$start_time,
                   waiting_time = .data$flow_time - .data$activity_time)
 
-  dispatch_next(metric, monitor_data, ...)
+  dispatch_next(metric, x)
 }
 
-plot_arrivals_activity_time <- function(monitor_data) {
-  ggplot(monitor_data) +
+plot.arrivals.activity_time <- function(x) {
+  ggplot(x) +
     aes_(x = ~end_time, y = ~activity_time) +
-    geom_line(aes_(group = ~replication), alpha = set_alpha(monitor_data)) +
+    geom_line(aes_(group = ~replication), alpha = set_alpha(x)) +
     stat_smooth() +
     xlab("simulation time") +
     ylab("activity time") +
@@ -23,10 +32,10 @@ plot_arrivals_activity_time <- function(monitor_data) {
     expand_limits(y = 0)
 }
 
-plot_arrivals_waiting_time <- function(monitor_data) {
-  ggplot(monitor_data) +
+plot.arrivals.waiting_time <- function(x) {
+  ggplot(x) +
     aes_(x = ~end_time, y = ~waiting_time) +
-    geom_line(aes_(group = ~replication), alpha = set_alpha(monitor_data)) +
+    geom_line(aes_(group = ~replication), alpha = set_alpha(x)) +
     stat_smooth() +
     xlab("simulation time") +
     ylab("waiting time") +
@@ -34,10 +43,10 @@ plot_arrivals_waiting_time <- function(monitor_data) {
     expand_limits(y = 0)
 }
 
-plot_arrivals_flow_time <- function(monitor_data) {
-  ggplot(monitor_data) +
+plot.arrivals.flow_time <- function(x) {
+  ggplot(x) +
     aes_(x = ~end_time, y = ~flow_time) +
-    geom_line(aes_(group = ~replication), alpha = set_alpha(monitor_data)) +
+    geom_line(aes_(group = ~replication), alpha = set_alpha(x)) +
     stat_smooth() +
     xlab("simulation time") +
     ylab("flow time") +