+++ R/plot.trajectory.R
@@ -1,4 +1,4 @@
-#' Plot method for trajectory objects
+#' Plot Method for \code{trajectory} Objects
 #'
 #' A method for the \code{\link{plot}} generic that plots a diagram of the given trajectory.
 #'